feat(ai): move ai from ee to main admin
