+++ packages/core/admin/ee/server/src/index.ts
@@ -10,10 +10,7 @@ import auditLogsController from './audit-logs/controllers/audit-logs';
 import { createAuditLogsService } from './audit-logs/services/audit-logs';
 import { createAuditLogsLifecycleService } from './audit-logs/services/lifecycles';
 import { auditLog } from './audit-logs/content-types/audit-log';
-import aiRoutes from './ai/routes/ai';
-import aiController from './ai/controllers/ai';
 import type { Core } from '@strapi/types';
-import { createAIAdminService } from './ai/containers/ai';
 
 const getAdminEE = () => {
   const eeAdmin = {
@@ -38,19 +35,14 @@ const getAdminEE = () => {
     controllers: {
       ...eeAdmin.controllers,
       ...(isAuditLogsEnabled ? { 'audit-logs': auditLogsController } : {}),
-      ai: aiController,
     },
     routes: {
       ...eeAdmin.routes,
       ...(isAuditLogsEnabled ? { 'audit-logs': auditLogsRoutes } : {}),
-      ai: aiRoutes,
     },
     async register({ strapi }: { strapi: Core.Strapi }) {
       // Run the default registration
       await eeAdmin.register({ strapi });
-
-      // Register the real admin service; isEnabled() on the service handles the guard
-      strapi.add('ai.admin', () => createAIAdminService({ strapi }));
 
       if (isAuditLogsEnabled) {
         // Register an internal audit logs service
