+++ packages/core/admin/server/src/ai/controllers/__tests__/ai.test.ts
@@ -1,6 +1,6 @@
 // @ts-expect-error - types are not generated for this file
 // eslint-disable-next-line import/no-relative-packages
-import createContext from '../../../../../../../../../tests/helpers/create-context';
+import createContext from '../../../../../../../../tests/helpers/create-context';
 import aiController from '../ai';
 
 describe('AI Controller', () => {
