+++ packages/core/admin/server/src/ai/routes/ai.ts
