+++ packages/core/admin/server/src/register.ts
@@ -2,6 +2,7 @@ import type { Core } from '@strapi/types';
 import registerAdminPanelRoute from './routes/serve-admin-panel';
 import adminAuthStrategy from './strategies/admin';
 import apiTokenAuthStrategy from './strategies/api-token';
+import { createAIAdminService } from './ai/containers/ai';
 
 export default ({ strapi }: { strapi: Core.Strapi }) => {
   const passportMiddleware = strapi.service('admin::passport').init();
@@ -9,6 +10,8 @@ export default ({ strapi }: { strapi: Core.Strapi }) => {
   strapi.server.api('admin').use(passportMiddleware);
   strapi.get('auth').register('admin', adminAuthStrategy);
   strapi.get('auth').register('content-api', apiTokenAuthStrategy);
+
+  strapi.add('ai.admin', () => createAIAdminService({ strapi }));
 
   const shouldServeAdminPanel = strapi.config.get('admin.serveAdminPanel');
